requirements.txt
# AI-Based Video Frame Interpolation System
# Python 3.10+

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Video processing
moviepy>=1.0.3

# Web interface
streamlit>=1.28.0

# Image processing and metrics
scikit-image>=0.21.0
matplotlib>=3.7.0

# Visualization
plotly>=5.17.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

